Apply useDeferredValue to stations in map
... and memoize the markers derived from it, as instructed in the React
documentation. This makes marker updates lower priority than other updates,
with the consequence that the UI does not freeze while they are updating.
This is most easily seen in how the selected station count and the filter
controls update (e.g., click All or None and the filter selector updates
before the markers are re-rendered. You can also interact with the filters
while the markers are updating (admittedly a bit tricky), and if you are
quick you do not see an intermediate marker update.